stocks_products/requirements.txt
django>=3.2
djangorestframework>=3.12
psycopg2-binary>=2.9
gunicorn>=21.0
